+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # 0.0.0 - 23-07-23
+## Updated
+- Creating the graph in `analysis.py`.
+
+# 0.0.0 - 23-07-23
 ## Created
 - Main file `analysis.py` where the network will be analyzed.
 - Supplementary file `common_utils.py` where generic functions will be added.
 - Folder `dataset` containing the datasets for the analysis.
 ## Added
 - `csv_to_df` and `convert_to_int` functions in `common_utils.py`.
+### TODO
+- Try commit `token_transfers_full.csv` large file whit better internet connection.